+++ noodles-bam/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Fixed
 
-  * bam/reader/header: Handle optional trailing `NUL` in SAM header text.
+  * bam/reader/header: Handle optional `NUL` padding in SAM header text.
 
 ## 0.49.0 - 2023-10-26